# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TaskFlow API
  description: Multi-tenant project management backend with PostgreSQL, Redis, BullMQ, and RBAC.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local Development Server (Default Port 8000)
  - url: http://localhost:3000
    description: Alternative Local Server (Port 3000)
paths:
...
  /projects:
    get:
      summary: List projects for active organization
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Projects list
    post:
      summary: Create new project
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: Website Redesign 2026
                description:
                  type: string
                  example: Redesigning corporate web app
      responses:
        '201':
          description: Project created
...
  /tasks:
    get:
      summary: List tasks with filters, pagination, and Full-Text Search
      security:
        - BearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [todo, in_progress, review, done]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: assignee
          in: query
          schema:
            type: string
            format: uuid
        - name: due_from
          in: query
          schema:
            type: string
            format: date-time
        - name: due_to
          in: query
          schema:
            type: string
            format: date-time
        - name: q
          in: query
          description: Full-Text Search on title and description
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: paginationType
          in: query
          schema:
            type: string
            enum: [offset, cursor]
            default: offset
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Tasks list
    post:
      summary: Create new task
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, title]
              properties:
                projectId:
                  type: string
                  format: uuid
                title:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [todo, in_progress, review, done]
                priority:
                  type: string
                  enum: [low, medium, high, urgent]
                dueDate:
                  type: string
                  format: date-time
                assigneeIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '201':
          description: Task created
...
  /tasks/{taskId}/comments:
    get:
      summary: List comments for task
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Comments list
    post:
      summary: Add comment to task
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
      responses:
        '201':
          description: Comment created
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT